+++ src/main/java/com/challenge/entrypoint/kafka/events/CreateOrderItemEvent.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record CreateOrderItemEvent(
         @NotBlank()
         String productUuid,
@@ -12,6 +14,6 @@ public record CreateOrderItemEvent(
         Integer quantity,
 
         @Positive()
-        Double price
+        BigDecimal price
 ) {
 }
